requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
whatsapp-api-client-python==0.0.46
openai>=1.50.0
chromadb==0.5.0
numpy<2
python-docx==1.1.0
google-api-python-client==2.140.0
google-auth==2.27.0
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.9.0

# Excel обработка
pandas==2.1.4
openpyxl==3.1.2

# Scheduler
apscheduler==3.10.4
